scaffolding for LearningLoop Sprint 1
I have put comments in every file for what I feel should be in there, as well as example JSONC and HTML files for how I imagine things implemented. This is the "scaffolding" we discussed during sprint planning. Feel free to change things as you see fit for your tasks! Leave a message in the Discord of any major changes.